Move ComparisonType into its own file
